+++ ospy/helpers.py
@@ -29,6 +29,13 @@ def now():
     return time.time() + (datetime.datetime.now() - datetime.datetime.utcnow()).total_seconds()
 
 
+def try_float(val, default=0):
+    try:
+        return float(val)
+    except ValueError:
+        return default
+
+
 def datetime_string(timestamp=None):
     if timestamp:
         if hasattr(timestamp, 'strftime'):
@@ -182,7 +189,7 @@ def uptime():
 
 
 def get_ip():
-    """Returns the IP adress if available."""
+    """Returns the IP address if available."""
     try:
         import subprocess
         arg = 'ip route list'
@@ -196,7 +203,7 @@ def get_ip():
 
 
 def get_mac():
-    """Retrun MAC from file"""
+    """Return MAC from file"""
     try:
         return str(open('/sys/class/net/eth0/address').read())
     except Exception: