+++ ospy/options.py
@@ -42,6 +42,22 @@ class _Options(object):
             "category": "System"
         },
         {
+            "key": "elevation",
+            "name": "Elevation (m)",
+            "default": 0,
+            "help": "Elevation of this location in meters.",
+            "category": "System",
+            "min": 0,
+            "max": 10000
+        },
+        {
+            "key": "wunderground_key",
+            "name": "Wunderground API key",
+            "default": "",
+            "help": "To make use of local weather conditions, a weather underground API key is needed.",
+            "category": "System"
+        },
+        {
             "key": "time_format",
             "name": "24-hour clock",
             "default": True,