+++ ospy/server.py
@@ -9,6 +9,8 @@ import os
 # Local imports
 from ospy.options import options
 from ospy.scheduler import scheduler
+from ospy.weather import weather
+from ospy.log import log
 
 import plugins