+++ ospy/stations.py
@@ -21,6 +21,9 @@ class _Station(object):
         self.enabled = True
         self.ignore_rain = False
         self.usage = 1.0
+        self.precipitation = 10.0
+        self.last_balance_date = datetime.date.today() - datetime.timedelta(days=21)
+        self.last_balance = 0.0
 
         # Remove (old) master info:
         if options.cls_name(self, index) in options: